requirements.txt
django>=6.0
gunicorn
psycopg2-binary
django-modeltranslation
pillow
